png transparency fix + clean published images + hasImages remove multiple images

+++ src/Commands/CleanCommand.php
@@ -9,7 +9,7 @@ use \App\ImageStore;
 class CleanCommand extends Command
 {
     use ConfirmableTrait;
-    protected $signature = 'colortools:clean {--delete}';
+    protected $signature = 'colortools:clean {--delete} {--deletePublished}';
     protected $description = 'Spring cleaning';
 
     public function __construct()
@@ -39,9 +39,13 @@ class CleanCommand extends Command
 
         $this->comment(count($extraFilesStored).' extra images found in storage '.str_plural('image', count($extraFilesStored)).' found with a total size of '.number_format(array_sum($extraFilesStored)/1024/1024, 2).'MB');
 
+        $filesPublished = [];
         $extraFilesPublished = [];
         foreach(glob($publicPath.'/*/*') as $publishedFile) {
             $hash = substr($publishedFile, 1 + strrpos($publishedFile, '/'), 32);
+            if($this->option('deletePublished', false)) {
+                $filesPublished[] = $publishedFile;
+            }
             if(in_array($hash, $storedHashes)) {
                 continue;
             }
@@ -51,13 +55,27 @@ class CleanCommand extends Command
 
         $this->comment(count($extraFilesPublished).' extra published '.str_plural('image', count($extraFilesPublished)).' found with a total size of '.number_format(array_sum($extraFilesPublished)/1024/1024, 2).'MB');
 
+        if($this->option('deletePublished', false)) {
+            $this->error('Deleting all published files - '.count($filesPublished).' '.str_plural('image', count($filesPublished)));
+            $deletedSize = 0;
+            foreach($filesPublished as $file) {
+                if(file_exists($file)) {
+                    $deletedSize += filesize($file);
+                    unlink($file);
+                }
+            }
+            $this->error('Freed up '.number_format($deletedSize/1024/1024, 2).'MB');
+        }
+
         if($this->option('delete', false)) {
             if(isset($toDelete)) {
                 $this->error('Deleting '.count($toDelete).' '.str_plural('image', count($toDelete)));
                 $deletedSize = 0;
                 foreach($toDelete as $file) {
-                    $deletedSize += filesize($file);
-                    unlink($file);
+                    if(file_exists($file)) {
+                        $deletedSize += filesize($file);
+                        unlink($file);
+                    }
                 }
                 $this->error('Freed up '.number_format($deletedSize/1024/1024, 2).'MB');
             } else {
@@ -65,7 +83,9 @@ class CleanCommand extends Command
             }
         } else {
             if(isset($toDelete)) {
-                $this->comment('Run php artisan colortools:clean --delete if you want to delete the extra files');
+                $this->comment('Run php artisan colortools:clean --delete if you want to delete the extra files or --deletePublished to clean the published images');
+            } else if (!$this->option('deletePublished', false)) {
+                $this->comment('Run php artisan colortools:clean --deletePublished to clean the published images');
             }
         }
